+++ mycodo/mycodo_flask/forms/forms_settings.py
@@ -396,3 +396,5 @@ class SettingsDiagnostic(FlaskForm):
     delete_maths = SubmitField(lazy_gettext('Delete All Maths'))
     delete_notes_tags = SubmitField(lazy_gettext('Delete All Notes and Note Tags'))
     delete_outputs = SubmitField(lazy_gettext('Delete All Outputs'))
+    delete_file_dependency = SubmitField(lazy_gettext('Delete File') + ': .dependency')
+    delete_file_upgrade = SubmitField(lazy_gettext('Delete File') + ': .upgrade')