+++ mycodo/mycodo_flask/routes_settings.py
@@ -355,14 +355,18 @@ def settings_diagnostic():
 
         if form_settings_diagnostic.delete_dashboard_elements.data:
             utils_settings.settings_diagnostic_delete_dashboard_elements()
-        if form_settings_diagnostic.delete_inputs.data:
+        elif form_settings_diagnostic.delete_inputs.data:
             utils_settings.settings_diagnostic_delete_inputs()
-        if form_settings_diagnostic.delete_maths.data:
+        elif form_settings_diagnostic.delete_maths.data:
             utils_settings.settings_diagnostic_delete_maths()
-        if form_settings_diagnostic.delete_notes_tags.data:
+        elif form_settings_diagnostic.delete_notes_tags.data:
             utils_settings.settings_diagnostic_delete_notes_tags()
-        if form_settings_diagnostic.delete_outputs.data:
+        elif form_settings_diagnostic.delete_outputs.data:
             utils_settings.settings_diagnostic_delete_outputs()
+        elif form_settings_diagnostic.delete_file_dependency.data:
+            utils_settings.settings_diagnostic_delete_file('dependency')
+        elif form_settings_diagnostic.delete_file_upgrade.data:
+            utils_settings.settings_diagnostic_delete_file('upgrade')
 
         return redirect(url_for('routes_settings.settings_diagnostic'))